==> Development/80 Replenishment Prep.ipynb <==
# %load D:/BUDS_MID/__init__.py
"""
Created on Wed Feb 19 16:21:00 2020

@author: cmchico

usual imports
"""

import os
import pandas as pd
import numpy as np

pd.options.display.max_rows = 100
pd.set_option('display.float_format', '{:.0f}'.format)

from time import time
from datetime import timedelta

import pickle

from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

path = 'D:/BUDS/ATM Recon/G4S Replenishment Data/'
out = 'D:/BUDS_MID/OJT ATM/ATM Forecasting Data New/'

start = time()
raw= pd.concat([pd.read_excel(path+i).assign(file=i) for i in os.listdir(path)], ignore_index=True)
raw['mon'] = raw.file.str.split(' ').str[3]
print(timedelta(seconds=time()-start))

raw=raw[~raw['Location Id'].isnull()].copy()

raw.mon.value_counts()

raw['mon_num2'] = raw.mon.map({'Jan':1,'Feb':2,'Mar':3,'Apr':4,'May':5,'Jun':6,'Jul':7,'Aug':8,'Sep':9,
                              'Oct':10,'Nov':11,'Dec':12})

raw['year_num'] = raw['Job Date'].str.split('/').str[2].astype(int)
raw['day_num'] = raw['Job Date'].str.split('/').str[0].astype(int)
raw['mon_num'] = raw['Job Date'].str.split('/').str[1].astype(int)

raw['Date'] = pd.to_datetime(raw['Job Date'])

select = raw.mon_num == raw.mon_num2
raw.loc[select,'Date'] = pd.to_datetime(raw.loc[select,'year_num'].astype(str)+'/'
                                        +raw.loc[select,'mon_num'].astype(str)+'/'
                                        +raw.loc[select,'day_num'].astype(str))
raw['Job Date'] = raw.Date

df = raw[['Location Id','Job Date','Job Time',
    'Last Load Amount','Current Load Amount','G4S Branch','G$S Branch','% Cash Return/Atm','Replen','Valid Cash Return',
    'Pre-Emptive','Remarks','Site Type','CATEGORY','PREMIUM SITES',
   'Atm Location']].copy()

df.dropna(subset=['Location Id'], inplace=True)
df['G4S Branch'] = df['G4S Branch'].fillna(df['G$S Branch'])
df = df.reset_index(drop=True)
df.drop(columns=['G$S Branch'], inplace=True)

df['_forecasted'] = df.Remarks.fillna("").str.contains('CASTED')
df['_cassettefault'] = df.Remarks.str.contains('XCF') | df.Remarks.str.contains('FAULT')
df['_bpi'] = df['Pre-Emptive'].astype(str).fillna('_').str.contains('BPI')
df['_others'] = (df._forecasted | df._bpi | df._cassettefault) == False
df[['_forecasted','_cassettefault','_bpi','_others']].sum()

remarks = pd.read_csv(out+'80 Review Replenishment Remarks.csv')

df = df.merge(remarks[['Remarks','FAILURE','FORECAST','BPI','OTHERS']].drop_duplicates(subset='Remarks'))

df['reload_type'] = 'FAILURE'
df.loc[~df.FORECAST.isnull(),'reload_type']="FORECAST"
df.loc[~df.OTHERS.isnull(),'reload_type']="OTHERS"
df.loc[~df.BPI.isnull(),'reload_type']="OTHERS"

df.reload_type.value_counts()

# df[~(df['Pre-Emptive'].astype(str).str.contains('PRE').values)& (df._cassettefault)].Remarks.unique()

# df[df._others].Remarks.unique().tolist()

# df[['Job Date','Replen','Valid Cash Return','Pre-Emptive','Remarks']].drop_duplicates().to_csv(out+'80 Review Replenishment Remarks.csv')

#retain PRE-EMPTIVE DUE TO CF in scenario analysis
# df.groupby(['Replen','Pre-Emptive']).size().reset_index()

#only 1 ATM changed location
# df[['Location Id','Atm Location']].drop_duplicates().groupby(['Location Id']).count()['Atm Location'].value_counts()

fin = df[['Location Id','Job Date','Last Load Amount','Current Load Amount','% Cash Return/Atm','Replen','Pre-Emptive','_forecasted','_cassettefault','_bpi','_others','reload_type']].copy()

fin.info()

fin['TID'] = fin['Location Id'].astype(int).astype(str)

fin['DATE'] = pd.to_datetime(fin['Job Date'], dayfirst=True)

fin.drop(columns=['Location Id','Job Date']).to_pickle(out+'Actual/80 Replenishment.pkl')
