==> Development/83 Prophet Forecast.ipynb <==
import numpy as np
import pandas as pd
import itertools
# import matplotlib.pyplot as plt

from datetime import timedelta
from time import time

import pickle
import os

from dask.distributed import Client, progress
import dask.bag as db

# import edslab_prophet_dask as edd
# import edslab_prophet as ed

pd.options.display.float_format = '{:.2f}'.format
pd.set_option('display.max_rows', 100)


from fbprophet import Prophet

client = Client(processes = True,n_workers=12)
client

path = 'D:/BUDS_MID/OJT ATM/ATM Forecasting Data New/'
HP = pd.read_csv(path + 'Actual/Holidays_ALL.csv')

old_rmse = pd.read_csv('D:/BUDS_MID/OJT ATM/ATM Forecasting Data/Prophet best model.csv', usecols=['atm','model','rmse'])
old_rmse.columns=['tid','old_model','old_rmse']
old_rmse.tid = old_rmse.tid.astype('str')

old_rmse

rmse = pd.concat([pd.read_csv(path+'/Prophet Train/'+file, usecols=['tid','model','dev_rmse','val_rmse'],
                             dtype={'tid':'str'}) for file in os.listdir(path+'/Prophet Train/')],ignore_index=True)
rmse['mean_rmse'] = rmse[['dev_rmse','val_rmse']].mean(axis=1) 
rmse.info()

# rmse.merge(old_rmse,how='left').to_csv(path+'/83 Train RMSE and old.csv')

best = rmse.groupby(['tid']).val_rmse.apply(pd.Series.argmin).map(rmse[:18].model.to_dict()).reset_index()

best

best.columns=['tid','best_model']
best.tid = best.tid.astype(str)
best['week'] = best.best_model.str[2:3].astype(int)
best['mon'] = best.best_model.str[3:].astype(int)
best.best_model.value_counts()

best.info()

rmse.head()

best = best.merge(rmse, left_on=['tid','best_model'], right_on=['tid','model'])

best.to_csv(path+'89 Best Model.csv')

atms = best.tid.unique()
len(atms)

data = pd.read_pickle(path+'Actual/82 Aggregated TIDBR WDL.pkl')
data.info()


data_train = data[data.DATE <= '2019-12-31'][['DATE','TID_BR','WDL_AMT_FIN']].copy()
data_train.columns=['ds','tid','y']
data_train

week = [0,2,3]
mon = [0,2,3,4,5,6]
hol = ['HP']
holidays=HP

scenario =  pd.DataFrame(list(itertools.product(hol,week,mon)),columns=['hol','week','mon'])
scenario['model'] = scenario.hol + scenario.week.astype('str') +scenario.mon.astype(str)
scenario['hol_num'] = scenario['hol'].map({'HP':0,'OP':1})

print('\ntotal scenario: {}'.format(scenario.shape[0]))

dev_length = sum(data_train.ds.dt.date.unique() < pd.to_datetime('20181231'))
print(dev_length)
period = '1 days'
horizon = 14

from fbprophet import Prophet

def train(series,val_dates,holidays,weekly=0,monthly=0):
    series = series.copy()
    val_dates = val_dates.copy()
    #define model
    modelf = Prophet(interval_width = 0.95,
                    daily_seasonality = False,
                    weekly_seasonality = False,
                    yearly_seasonality = False,
                    holidays = holidays)
    if weekly != 0: modelf.add_seasonality(name='weeklyx', period = 7, fourier_order = weekly)
    if monthly != 0: modelf.add_seasonality(name='monthlyx', period = 30.5, fourier_order = monthly)
    
    modelf.fit(series)
    
#     return modelf
    val_dates['yhat'] = modelf.predict(val_dates[['ds']]).yhat.values.clip(min=0)
    val_dates['pred_date'] = series.tail(1).ds.dt.date.values.tolist()[0]
    val_dates['pred_day'] = ['day '+str(i) for i in range(val_dates.shape[0])]
    return val_dates

for atm in atms:
    start=time()

    series = data_train[data_train.tid == atm][['ds','y']].copy().reset_index(drop=True)
    index = best[best.tid == atm].index
    week = best.week[index].tolist()[0]
    mon = best.mon[index].tolist()[0]
    n=series.shape[0]-dev_length-horizon-1

    dbseries = db.from_sequence([series[i:i+dev_length] for i in range(n)])
    dbforecast = db.from_sequence([series[i+dev_length:i+dev_length+horizon][['ds','y']] for i in range(n)])

    dbmaster = db.map(train,dbseries,dbforecast,HP,week,mon)


    model = dbmaster.compute()
    model=pd.concat([model[i] for i in range(len(model))])
    model['se'] = (model.y-model.yhat)**2
    model['tid'] = atm
    rmse = model.groupby([model.ds.dt.month,'pred_day']).se.mean().agg(np.sqrt).reset_index()
    model.to_csv(path+'Prophet Forecast/'+atm+'_forecast.csv')
    rmse.to_csv(path+'Prophet Forecast RMSE/'+atm+'_forecast_rmse.csv')
    print(f'\nATM:{atm}  RMSE:{rmse.se.mean()}')
    print(timedelta(seconds=time()-start))

### Get best model
