==> Development/84 Optimization Test Thresh Recomm.ipynb <==
# %load D:/BUDS_MID/__init__.py
"""
Created on Wed Feb 19 16:21:00 2020

@author: cmchico

usual imports
"""

import os
import pandas as pd
import numpy as np

pd.options.display.max_rows = 100
pd.set_option('display.float_format', '{:.0f}'.format)


import f_validate_g4s as ed

from time import time
from datetime import timedelta

import pickle

from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

# %load D:/BUDS_MID/__initdask__.py
"""
Created on Wed Feb 19 16:24:03 2020

@author: cmchico
"""
import dask.bag as db
from dask.distributed import Client, progress

client = Client(processes = True,n_workers=12)
client

path = "D:/BUDS_MID/OJT ATM/ATM Forecasting Data New/"

thresh = pd.read_pickle(path+'89 Threshold Recommendations.pkl')
thresh = thresh[['tid','thresh']].drop_duplicates()
thresh.tid = thresh.tid.astype(str)
thresh.info()

test_dates = pd.date_range(start='01/01/2019',end='12/31/2019')

mapp = pd.read_pickle(path+'Actual/89 TID Br map.pkl')
#211 branches with just one ATM, can be treated as an ATM i/o Branch
mapp[mapp.TID!=mapp.TID_BR].BR_TIDCOUNT.value_counts()

reload_df = pd.read_pickle(path+'Actual/80 Replenishment.pkl')
reload_df = reload_df.sort_values(['TID','DATE'],ignore_index=True)
reload_df['Pre-Emptive'] = reload_df['Pre-Emptive'].str.strip().fillna("NORMAL LOADING")
reload_df['date_str'] = reload_df.DATE.astype(str)


reload_df['Current Load Amount'] = reload_df['Current Load Amount'].astype(str).str.strip().str.replace("","")
select = reload_df['Current Load Amount']==""
reload_df.loc[select,'Current Load Amount'] = "0"
reload_df['Current Load Amount'] = reload_df['Current Load Amount'].astype(int)

reload_df = reload_df[reload_df.DATE.isin(test_dates)].copy()
reload_df = reload_df.sort_values(['TID','DATE'],ignore_index=True)
reload_df['TID_BR'] = reload_df.TID.map(mapp.set_index('TID').TID_BR.to_dict())

first_reload = reload_df.drop_duplicates(subset=(['TID'])).copy()

reload_df.info()

reload_df.TID_BR.nunique()

reload_df.reload_type.value_counts()

preempt = reload_df[reload_df.reload_type.isin(['FAILURE','OTHERS'])][['date_str','TID_BR','reload_type','Current Load Amount']]

df = pd.concat([pd.read_csv(path+'Prophet Forecast/'+i, dtype={'ds':'str','pred_date':'str','tid':'str'}) 
                for i in os.listdir(path+'Prophet Forecast/')],ignore_index=True)
df['pred_date_D'] = pd.to_datetime(df.pred_date)
df.info()

df.tid.nunique()

df.sort_values(['tid','pred_date','ds'], inplace=True)

actual = df[df.pred_day.isin(['day 0','day 1','day 2'])].pivot_table(index=['tid','pred_date','pred_date_D'],columns='pred_day',values='yhat',aggfunc=sum).reset_index()
actual.columns=['tid','pred_date','pred_date_D','yhat0','yhat1','yhat2']
actual[['date_str','y']] = df[df.pred_day=='day 0'][['ds','y']].reset_index(drop=True)

actual.tid.nunique()

atms = [atm for atm in actual.tid.unique() if atm in mapp[mapp.BR_TIDCOUNT==1].TID_BR.unique()]
len(atms)

def par(atm,thresh,rel_amt):
    
    date_o = first_reload.loc[first_reload['TID_BR'] == atm].reset_index().DATE[0]
    actual_atm = actual[(actual.tid== atm) & (actual.pred_date_D >= date_o-timedelta(days=1))].reset_index(drop=True).copy()

    g4s=preempt[preempt.TID_BR==atm].set_index('date_str')
    g4s['Current Load Amount'] = rel_amt
    g4s['Pre-Emptive'] = g4s.reload_type
    
    threshold = thresh/rel_amt
    if g4s.shape[0]==0: g4s=None
    sched_= ed.sched(actual_atm,rel_amt=rel_amt,pred=pred,r=r,reload_charge=reload_charge,
                      threshold=threshold, hor=hor,g4s=g4s)
    sched_['fix_rel']=rel_amt
    sched_['thresh'] = thresh
    sched_['thresh%'] = threshold
    ss=sched_.groupby('cog_preemptive') \
                        [['reload_flag','cashout_flag','thresh_breach_flag','necessary_flag',
                                             'civ_cost','reload_cost']].sum().reset_index()
    ss['tid']=atm
    ss['fix_rel'] = rel_amt
    ss['thresh'] = thresh
    ss['thresh%'] = threshold
    sched_.to_csv(path+'Prophet Test Thresh R/_sched_'+atm+'_'+str(rel_amt)[:2]+'.csv')
    ss.to_csv(path+'Prophet Test Perf Thresh R/_schedperf_'+ atm+'_'+str(rel_amt)[:2]+'.csv')
    
    return sched_,ss

rel_amt=3200000
pred = ['yhat0','yhat1','yhat2']
hor=len(pred)-1
r = 0.05/365
reload_charge=2000
# threshold=0.25

#3atms per minute
# rel_amt = 3200000
# start=time()
# for atm in atms[:10]:
#     print(atm)
#     thresh_ = thresh[thresh.tid==atm].reset_index(drop=True).thresh[0]
#     a,b= par(atm,thresh_,rel_amt)
#     print(timedelta(seconds=time()-start))

atm

atm=atms[0]
thresh_ = thresh[thresh.tid==atm].reset_index(drop=True).thresh[0]
a=par(atm,thresh_,3200000)
# # par(atm,thresh_,4200000)
# # par(atm,thresh_,5200000)
# # par(atm,thresh_,6200000)

a[1]

dbatms = db.from_sequence(atms)
dbthresh = db.from_sequence([thresh[thresh.tid==atm].reset_index(drop=True).thresh[0] for atm in atms])
dbmaster = dbatms.map(par,dbthresh,rel_amt=3200000)

start=time()
df = dbmaster.compute()
print(timedelta(seconds=time()-start))

dbatms = db.from_sequence(atms)
dbthresh = db.from_sequence([thresh[thresh.tid==atm].reset_index(drop=True).thresh[0] for atm in atms])
dbmaster = dbatms.map(par,dbthresh,rel_amt=4200000)

start=time()
df = dbmaster.compute()
print(timedelta(seconds=time()-start))

dbatms = db.from_sequence(atms)
dbthresh = db.from_sequence([thresh[thresh.tid==atm].reset_index(drop=True).thresh[0] for atm in atms])
dbmaster = dbatms.map(par,dbthresh,rel_amt=5200000)

start=time()
df = dbmaster.compute()
print(timedelta(seconds=time()-start))

dbatms = db.from_sequence(atms)
dbthresh = db.from_sequence([thresh[thresh.tid==atm].reset_index(drop=True).thresh[0] for atm in atms])
dbmaster = dbatms.map(par,dbthresh,rel_amt=6200000)

start=time()
df = dbmaster.compute()
print(timedelta(seconds=time()-start))

perf = pd.concat(pd.re)

perf = pd.concat(pd.read_csv(path+'Prophet Test Perf Thresh R/' + i).assign(file=i)
          for i in os.listdir(path+'Prophet Test Perf Thresh R')[2:])

perfs = perf.groupby(['tid','file'])[['civ_cost','reload_cost','cashout_flag','reload_flag']].sum().reset_index()
perfs['total_cost'] = perfs.civ_cost+perfs.reload_cost
perfs['reload_amt'] = perfs.file.str.split('_').str[3].str.split(".").str[0].astype(int)*100000

mcost = perfs.sort_values(['total_cost','cashout_flag']).drop_duplicates(subset='tid')

mcost.sum()
mcost.mean()

mcashout = perfs.sort_values(['cashout_flag','total_cost']).drop_duplicates(subset='tid')

mcashout.sum()
mcashout.mean()

g4s = pd.read_csv(path+'85 G4S perf 2019.csv', dtype={'tid':'str'}).merge(mapp,left_on='tid',right_on='TID')

g4ss = g4s[g4s.TID_BR.isin(atms)].groupby('TID_BR')[['civ_cost','reload_cost','cashout_flag','reload_flag']].sum().reset_index()
g4ss['total_cost'] = g4ss.civ_cost + g4ss.reload_cost

g4ss.TID_BR.nunique()

g4ss.sum()
g4ss.mean()

225337837-209006445

# len(atms)

# generated = pd.Series(os.listdir(path+'/Prophet Test Thresh R/OFFSITE OR ONE ATM BRANCH')).str.split('_').str[2].unique()

# len(generated)

# atms_new = list(set(atms) - set(generated))

# len(atms_new)

# dbatms = db.from_sequence(atms_new)
# dbthresh = db.from_sequence([thresh[thresh.tid==atm].reset_index(drop=True).thresh[0] for atm in atms_new])
# dbmaster = dbatms.map(par,dbthresh,rel_amt=3200000)

# start=time()
# df = dbmaster.compute()
# print(timedelta(seconds=time()-start))

# dbatms = db.from_sequence(atms_new)
# dbthresh = db.from_sequence([thresh[thresh.tid==atm].reset_index(drop=True).thresh[0] for atm in atms_new])
# dbmaster = dbatms.map(par,dbthresh,rel_amt=4200000)

# start=time()
# df = dbmaster.compute()
# print(timedelta(seconds=time()-start))

# dbatms = db.from_sequence(atms_new)
# dbthresh = db.from_sequence([thresh[thresh.tid==atm].reset_index(drop=True).thresh[0] for atm in atms_new])
# dbmaster = dbatms.map(par,dbthresh,rel_amt=5200000)

# start=time()
# df = dbmaster.compute()
# print(timedelta(seconds=time()-start))

# dbatms = db.from_sequence(atms_new)
# dbthresh = db.from_sequence([thresh[thresh.tid==atm].reset_index(drop=True).thresh[0] for atm in atms_new])
# dbmaster = dbatms.map(par,dbthresh,rel_amt=6200000)

# start=time()
# df = dbmaster.compute()
# print(timedelta(seconds=time()-start))
