==> Development/85 G4S Scenario.ipynb <==
import numpy as np
import pandas as pd

import f_validate_g4s as ed

from datetime import timedelta
from time import time

pd.options.display.float_format = '{:.2f}'.format
pd.set_option('display.max_rows', 100)

# %load D:/BUDS_MID/__initdask__.py
"""
Created on Wed Feb 19 16:24:03 2020

@author: cmchico
"""
import dask.bag as db
from dask.distributed import Client, progress
from dask import delayed

client = Client(processes = True,n_workers=12)
client

path = 'D:/BUDS_MID/OJT ATM/ATM Forecasting Data New/'

test_dates = pd.date_range(start='01/01/2019',end='12/31/2019')

reload_df = pd.read_pickle(path+'Actual/80 Replenishment.pkl')
reload_df = reload_df.sort_values(['TID','DATE'],ignore_index=True)
reload_df['Pre-Emptive'] = reload_df['Pre-Emptive'].str.strip().fillna("NORMAL LOADING")
reload_df['date_str'] = reload_df.DATE.astype(str)


reload_df['Current Load Amount'] = reload_df['Current Load Amount'].astype(str).str.strip().str.replace("","")
select = reload_df['Current Load Amount']==""
reload_df.loc[select,'Current Load Amount'] = "0"
reload_df['Current Load Amount'] = reload_df['Current Load Amount'].astype(int)

reload_df.info()

reload_df['Pre-Emptive'].value_counts()

actual=pd.read_pickle(path+'Actual/82 Aggregated TID WDL.pkl')
actual['AMOUNT_SUM'] = actual['WDL_AMT']
actual['date_str'] = actual.DATE.astype(str)
actual = actual[actual.DATE.isin(test_dates)].copy()
actual.TID = actual.TID.astype(str)
actual.info()

reload_df = reload_df[reload_df.DATE.isin(test_dates)].copy()
reload_df = reload_df.sort_values(['TID','DATE'],ignore_index=True)

first_reload = reload_df.drop_duplicates(subset=(['TID'])).copy()

# actual
atms_raw = actual.TID.unique().tolist()
atms = [atm for atm in atms_raw if atm in first_reload.TID.unique()]

print(f'No. of ATMs from Raw: {len(atms_raw)}')

print(f'NO of ATMs from G4S: {first_reload.TID.nunique()}')
print(f'NO of ATMs for simulation: {len(atms)}')

r = 0.05/365
reload_charge=2000
threshold=0.25

def par(atm):
    date_o = first_reload.loc[first_reload['TID'] == atm].reset_index().DATE[0]
    actual_atm = actual[(actual['TID'] == atm) & (actual.DATE >= date_o)].reset_index(drop=True).copy()
    
    g4s = reload_df[reload_df['TID'] == atm].set_index('date_str').copy()
    
    g4s_all_sched = ed.validate_g4s(actual_atm, g4s, r=r,reload_charge=reload_charge, threshold=threshold)
    
    g4s_all_sched.to_csv(path+'G4S Schedule/_sched_'+ atm+'.csv')
    ss=g4s_all_sched.groupby('cog_preemptive') \
                        [['reload_flag','cashout_flag','thresh_breach_flag','necessary_flag',
                                             'civ_cost','reload_cost']].sum().reset_index()
    ss['tid']=atm
    
    ss.to_csv(path+'G4S Schedule Perf/_schedperf_'+ atm+'.csv')
    return g4s_all_sched,ss

g4s_all_sched = dict()
g4s_all_ssum = dict()

print(f'No. of ATMs from Raw: {len(atms)}')

for atm in atms[296:300]:
    #first reload date
    print(atm)
    start = time()
    date_o = first_reload.loc[first_reload['TID'] == atm].reset_index().DATE[0]
    actual_atm = actual[(actual['TID'] == atm) & (actual.DATE >= date_o)].reset_index(drop=True).copy()
    
    g4s = reload_df[reload_df['TID'] == atm].set_index('date_str').copy()
    
    g4s_all_sched[atm] = ed.validate_g4s(actual_atm, g4s, r=r,reload_charge=reload_charge, threshold=threshold)
    
    g4s_all_sched[atm].to_csv(path+'G4S Schedule/_sched_'+ atm+'.csv')
    ss=g4s_all_sched[atm].groupby('cog_preemptive') \
                        [['reload_flag','cashout_flag','thresh_breach_flag','necessary_flag',
                                             'civ_cost','reload_cost']].sum().reset_index()
    ss['tid']=atm
    
    ss.to_csv(path+'G4S Schedule Perf/_schedperf_'+ atm+'.csv')
    g4s_all_ssum[atm] = ss
print(timedelta(seconds=time()-start))

dbatms = db.from_sequence(atms)
dbmaster = dbatms.map(par)

start=time()
df = dbmaster.compute()
print(timedelta(seconds=time()-start))

g4s_perf = pd.concat([df[i][1] for i in range(len(atms))], ignore_index=True)

g4s_sched = pd.concat([df[i][0].assign(tid=atms[i]) for i in range(len(atms))], ignore_index=True)

g4s_perf.to_csv(path+'85 G4S perf 2019.csv')
g4s_sched.to_pickle(path+'85 G4S sched 2019.pkl')

g4s_sched
