==> Development/86 Optimization Test Branch.ipynb <==
# %load D:/BUDS_MID/__init__.py
"""
Created on Wed Feb 19 16:21:00 2020

@author: cmchico

usual imports
"""

import os
import pandas as pd
import numpy as np

pd.options.display.max_rows = 100
pd.set_option('display.float_format', '{:.0f}'.format)

import f_sched_branch as ed


from time import time
from datetime import timedelta

import pickle

from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

# %load D:/BUDS_MID/__initdask__.py
"""
Created on Wed Feb 19 16:24:03 2020

@author: cmchico
"""
import dask.bag as db
from dask.distributed import Client, progress

client = Client(processes = True,n_workers=12)
client

##### parameters and data source

path = "D:/BUDS_MID/OJT ATM/ATM Forecasting Data New/"

test_dates = pd.date_range(start='01/01/2019',end='12/31/2019')

r_mapp = pd.read_pickle(path+'Actual/89 TID Br map.pkl')
r_thresh = pd.read_pickle(path+'89 Threshold Recommendations.pkl')
r_reload_df = pd.read_pickle(path+'Actual/80 Replenishment.pkl')
r_df = pd.concat([pd.read_csv(path+'Prophet Forecast/'+i, dtype={'ds':'str','pred_date':'str','tid':'str'}) 
                for i in os.listdir(path+'Prophet Forecast/')],ignore_index=True)
r_actual = pd.read_pickle(path+'Actual/82 Aggregated TID WDL.pkl')

##### count of atm per group (branch)

#211 branches with just one ATM, can be treated as an ATM i/o Branch
mapp = r_mapp.copy()
atm_count_dict = mapp[['TID_BR','BR_TIDCOUNT']].drop_duplicates().set_index('TID_BR').BR_TIDCOUNT.to_dict()
atm_br_dict = mapp.set_index('TID').TID_BR.to_dict()

mapp[['TID_BR','BR_TIDCOUNT']].drop_duplicates().BR_TIDCOUNT.value_counts()

##### threshold

r_thresh

thresh = r_thresh.copy()
thresh['per_rmse_yave'] =thresh.rmse/thresh.y_ave*100
thresh.tid = thresh.tid.astype(str)
thresh['tid_br'] = thresh.tid
thresh = thresh.sort_values(['per_rmse_yave','pred_date'], ascending=False)
thresh['tid_count'] = thresh['tid'].map(atm_count_dict)
# thresh.info()

thresh[thresh.tid_count > 1]

atms = thresh[thresh.tid_count > 1].tid_br.unique()
len(atms)

##### replenishment

reload_df = r_reload_df.copy()
reload_df = reload_df.sort_values(['TID','DATE'],ignore_index=True)
reload_df['Pre-Emptive'] = reload_df['reload_type'].str.strip().fillna("NORMAL LOADING")
reload_df['date_str'] = reload_df.DATE.astype(str)

reload_df['Current Load Amount'] = reload_df['Current Load Amount'].astype(str).str.strip().str.replace("","")
select = reload_df['Current Load Amount']==""
reload_df.loc[select,'Current Load Amount'] = "0"
reload_df['Current Load Amount'] = reload_df['Current Load Amount'].astype(int)

reload_df = reload_df[reload_df.DATE.isin(test_dates)].copy()
reload_df = reload_df.sort_values(['TID','DATE'],ignore_index=True)
reload_df['tid'] = reload_df.TID
reload_df['tid_br'] = reload_df.TID.map(atm_br_dict)
reload_df['tid_count'] = reload_df.tid_br.map(atm_count_dict)
# reload_df[['_forecasted','_cassettefault','_bpi','_others']] = \
#         (reload_df[['_forecasted','_cassettefault','_bpi','_others']]*1).fillna(0)


reload_df =  reload_df[reload_df.tid_br.isin(atms)].reset_index(drop=True)
reload_df.info()

# type_reload = np.array(['forecasted','cassettefault','bpi','others'])
# def get_type(x):
#     return '-'.join((type_reload[x.values]).tolist())

# reload_df['reload_type'] = reload_df[['_forecasted','_cassettefault','_bpi','_others']].fillna(False) \
#                                 .apply(get_type, axis=1)

# reload_df.groupby(['Pre-Emptive', 'reload_type']).size()
# # reload_df.reload_type.value_counts(normalize=True)*100

first_together = reload_df.groupby(['tid_br','tid_count','DATE']).size().reset_index()
first_together.columns = ['tid_br','tid_count','DATE','count_date']
first_together['check'] =first_together.tid_count == first_together.count_date
first_reload= first_together[first_together.check].sort_values(['DATE']).drop_duplicates(subset='tid_br')
first_reload.DATE.dt.month.value_counts()

reload_df

preempt = reload_df[reload_df.reload_type != 'FORECAST'][['date_str','tid','tid_br','reload_type','Current Load Amount']] 
preempt

##### actual withdrawal

wdl = r_actual[r_actual.DATE.isin(test_dates)].copy()
wdl['tid'] = wdl.TID.astype(str)
wdl['tid_br'] = wdl.tid.map(atm_br_dict)
wdl['date_str'] = wdl.DATE.astype(str)
wdl = wdl[wdl.tid_br.isin(atms)].reset_index(drop=True)
wdl.info()

##### forecasts

df = r_df[r_df.tid.isin(atms)].copy()
df['tid_br'] = df.tid
df.sort_values(['tid','pred_date','ds'], inplace=True)
df['pred_date_D'] = pd.to_datetime(df.pred_date)
df.info()

actual = df.pivot_table(index=['tid_br','pred_date','pred_date_D'],columns='pred_day',values='yhat',aggfunc=sum).reset_index()
actual[['date_str','y']] = df[df.pred_day=='day 0'][['ds','y']].reset_index(drop=True)
actual = actual[actual.tid_br.isin(atms)].reset_index(drop=True)
actual.shape
actual.head()

#round-off forecasts to nearest 1000
actual[[x for x in actual.columns if 'day' in x]] = \
        (actual[[x for x in actual.columns if 'day' in x]] / 1000).astype(int)*1000

def data_prep(atm):
    atm_no = len(tid)
    date_o = first_reload[first_reload.tid_br == atm].reset_index().DATE[0]
    
    wdl_atm = wdl[wdl.tid_br==atm].pivot_table(index=['date_str','tid_br'], columns=['tid'],values=['WDL_AMT_FIN']).reset_index()
    wdl_atm.columns = wdl_atm.columns.get_level_values(0)[:2].tolist() + ('wdl_' + wdl_atm.columns.get_level_values(1)[2:]).tolist()


    actual_atm = actual[(actual.tid_br== atm) & (actual.pred_date_D >= date_o-timedelta(days=1))] \
                    [['tid_br','pred_date','date_str','y'] + window_col].reset_index(drop=True) \
                    .merge(wdl_atm,how='left')

    #in case more than 1 reload type per date, just get the first one to flag all atms
    g4s = preempt[preempt.tid_br==atm][['date_str','tid_br','reload_type']].drop_duplicates(subset=['date_str','tid_br']) \
        .merge(preempt[preempt.tid_br==atm][['date_str','tid_br','tid','Current Load Amount']])
    g4s = g4s.pivot_table(index=['date_str','tid_br','reload_type'],
                            columns=['tid'],values=['Current Load Amount']).reset_index().fillna(0)
    g4s.columns = g4s.columns.get_level_values(0)[:3].tolist() + ('c_'+g4s.columns.get_level_values(1)[3:]).tolist()
    
    return actual_atm, g4s

def perf(aa,reload_amt,atm_no,tid):
    
    ss =aa.groupby(['reload_type'])[['civ_cost','reload_cost','total_cost','cashout_flag','reload_flag']].sum().reset_index()
    reloads = {tid[i]+ str(i):reload_amt[i] for i in range(atm_no)}

    a=pd.Series(reloads).reset_index()
    a.columns=['tid','reload_amt']

    ss['tid_br']=atm
    ss['thresh'] = threshold
    ss['thresh%'] = thresh[thresh.tid_br==atm].per_rmse_yave.astype(int).unique()[0]

    avail = (aa[['avail_' +str(i) for i in range(atm_no)]].sum()/aa.shape[0]*100).astype(int).reset_index(drop=True)
    avail.name = 'Availability'

    ss=ss.join(a.join(avail),how='left')
    return ss

def par(reload_amt,i,actual_atm,atm, tid,g4s,t,threshold,r,reload_charge,days_no,hor):
    atm_no=len(tid)
    try:
        g4s[g4s.columns[3:]]= (g4s[g4s.columns[3:]]>0)*reload_amt
        aa = ed.sched_branch(actual_atm,reload_amt, tid,t,threshold,
                           r, reload_charge,days_no,hor,g4s)
    except:
        return [[atm,tid,reload_amt,'fail']]
    
    aa.to_csv(path+'Prophet Test Thresh R/'+'_'+atm+'_'+str(i)+'_.csv')
    ss = perf(aa,reload_amt,atm_no,tid)
    ss.to_csv(path+'Prophet Test Perf Thresh R/'+'_'+atm+'_'+str(i)+'_.csv')
    return [[atm,tid,reload_amt,'success']]
#     except:
        

def par_atm(atm):
    threshold = thresh[thresh.tid==atm].thresh.unique()[0]
    tid = wdl[wdl.tid_br== atm].tid.unique().tolist()
    atm_no=len(tid)    
    t = min(100000,int(threshold/hor/atm_no/100000)*100000)
    
    actual_atm, g4s = data_prep(atm)

    
    scenario = list(itertools.permutations(choice,atm_no))
    scenario_all = [list(i) for i in scenario] + [[a for i in range(atm_no)] for a in choice]
    
    print(len(scenario_all))
    error = [[atm,tid,scenario_all[i],"success"] for i in range(len(scenario_all))]
    
    for i in range(len(scenario_all)):
        reload_amt=scenario_all[i]
        try:
            par(reload_amt,i,actual_atm,atm, tid,g4s,t,threshold,r,reload_charge,days_no,hor)
            
#             g4s[g4s.columns[3:]]= (g4s[g4s.columns[3:]]>0)*reload_amt
#             aa=ed.sched_branch(actual_atm,reload_amt, tid,t,threshold,
#                                r, reload_charge,days_no,hor,g4s)
#             aa.to_pickle(path+'Prophet Test Thresh R/'+'_'+atm+'_'+str(i)+'_.pkl')
#             ss = perf(aa,atm_no,tid)
#             print(ss.total_cost.sum())
#             ss.to_csv(path+'Prophet Test Perf Thresh R/'+'_'+atm+'_'+str(i)+'_.csv')
        except:
            error[i] = [atm,tid,scenario_all[i],"fail"]
        print(timedelta(seconds=time()-start))
    return error


import itertools
choice=[3200000,4200000,5200000,6200000]


atm_no=2
scenario = list(itertools.permutations(choice,atm_no))
scenario_all = [list(i) for i in scenario] + [[a for i in range(atm_no)] for a in choice]
print(len(scenario_all))

atm_no=3
scenario = list(itertools.permutations(choice,atm_no))
scenario_all = [list(i) for i in scenario] + [[a for i in range(atm_no)] for a in choice]
print(len(scenario_all))
scenario_all
# atm_no=4
# scenario = list(itertools.permutations(choice,atm_no))
# scenario_all = [list(i) for i in scenario] + [[a for i in range(atm_no)] for a in choice]
# print(len(scenario_all))

scenario_all

##### start

hor=2
days_no = 7
r = 0.05/365
reload_charge=2000
window_col = ['day '+ str(i) for i in range(days_no+hor)]

import itertools

choice=[3200000,4200000,5200000,6200000]

scenario_ = pd.read_csv(path+'93 Scenarios.csv')
scen=dict()
scen[1] = [[x] for x in scenario_[scenario_.atm_no==1].rel1.values] 
scen[2] = scenario_[scenario_.atm_no==2][['rel1','rel2']].values.tolist()
scen[3] =  scenario_[scenario_.atm_no==3][['rel1','rel2','rel3']].values.tolist()
scen[4] = scenario_[scenario_.atm_no==4][['rel1','rel2','rel3','rel4']].values.tolist()

len(scen[4])

atm='BR003'
threshold = thresh[thresh.tid==atm].thresh.unique()[0]
tid = wdl[wdl.tid_br== atm].tid.unique().tolist()
atm_no=len(tid)    
t = min(100000,int(threshold/hor/atm_no/100000)*100000)

scenario_all=scen[atm_no]
print(len(scenario_all))

actual_atm, g4s = data_prep(atm)

# a=par(scenario_all[0],0,actual_atm=actual_atm,atm=atm, 
#            tid=tid,g4s=g4s,t=t,threshold=threshold,
#            r=r,reload_charge=reload_charge,days_no=days_no,hor=hor)

### sequential 2mins

start=time()
# for i in range(len(scenario_all)):
for i in range(1):
    reload_amt=scenario_all[i]
    g4s[g4s.columns[3:]]= (g4s[g4s.columns[3:]]>0)*reload_amt
    aa=ed.sched_branch(actual_atm,reload_amt, tid,t,threshold,
                       r, reload_charge,days_no,hor,g4s)
    aa.to_pickle(path+'Prophet Test Thresh R/'+'_'+atm+'_'+str(i)+'_.pkl')
    ss = perf(aa,reload_amt, atm_no,tid)
    print(ss.total_cost.sum())
    ss.to_csv(path+'Prophet Test Perf Thresh R/'+'_'+atm+'_'+str(i)+'_.csv')
    print(timedelta(seconds=time()-start))

### 583 PROBLEMATIC

### parallel scenario, one atm 1:08secs

atm_new = list(set(atms)-set(['BR583']))

atm_new[59:]
atm_new.index('BR056')

start = time()
for atm in newatm:
    print(atm)
    threshold = thresh[thresh.tid==atm].thresh.unique()[0]
    tid = wdl[wdl.tid_br== atm].tid.unique().tolist()
    atm_no=len(tid)    
    
    t = min(100000,int(threshold/hor/atm_no/100000)*100000)
    
    actual_atm, g4s = data_prep(atm)

    scenario = list(itertools.permutations(choice,atm_no))
    scenario_all = [list(i) for i in scenario] + [[a for ia in range(atm_no)] for a in choice]

    dbscen = db.from_sequence(scenario_all[i] for i in range(len(scenario_all)))
    dbi = db.from_sequence(i for i in range(len(scenario_all)))
    dbmaster = \
    db.map(par,dbscen,dbi,actual_atm=actual_atm,atm=atm, 
           tid=tid,g4s=g4s,t=t,threshold=threshold,
           r=r,reload_charge=reload_charge,days_no=days_no,hor=hor)
    a = dbmaster.compute()
    b = pd.concat([pd.DataFrame(a[i], columns=['atm','tid','reload_amt','status']) for i in range(len(scenario_all))]
          ,ignore_index=True)
    pd.DataFrame(b).to_pickle(path+'Prophet Forecast All/_'+atm+'_.pkl')
    print(b.status.value_counts())
    print(timedelta(seconds=time()-start))

for atm in ['BR347', 'BR377']:
    start = time()
    print(atm)
    threshold = thresh[thresh.tid==atm].thresh.unique()[0]
    tid = wdl[wdl.tid_br== atm].tid.unique().tolist()
    atm_no=len(tid)    
    
    t = min(100000,int(threshold/hor/atm_no/100000)*100000)
    
    actual_atm, g4s = data_prep(atm)

    scenario = list(itertools.permutations(choice,atm_no))
    scenario_all = [list(i) for i in scenario] + [[a for ia in range(atm_no)] for a in choice]

    dbscen = db.from_sequence(scenario_all[i] for i in range(len(scenario_all)))
    dbi = db.from_sequence(i for i in range(len(scenario_all)))
    dbmaster = \
    db.map(par,dbscen,dbi,actual_atm=actual_atm,atm=atm, 
           tid=tid,g4s=g4s,t=t,threshold=threshold,
           r=r,reload_charge=reload_charge,days_no=days_no,hor=hor)
    a = dbmaster.compute()
    b = pd.concat([pd.DataFrame(a[i], columns=['atm','tid','reload_amt','status']) for i in range(len(scenario_all))]
          ,ignore_index=True)
    pd.DataFrame(b).to_pickle(path+'Prophet Forecast All/_'+atm+'_.pkl')
    print(b.status.value_counts())
    print(timedelta(seconds=time()-start))

generated = pd.Series(os.listdir(path+'Prophet Test Thresh R/'))
# .str.split('_').str[1].unique()

generated = generated[generated.str.contains("csv")]
generated = generated.str.split('_').str[1].unique()

generated

newatm = list(set(atms)-set(generated))
newatm

os.listdir(path+'Prophet Test Thresh R/')

### parallel atm, sequential scenario

#per atm
start=time()
dbatm = db.from_sequence(atms)
dbmaster = db.map(par_atm,dbatm)
a=dbmaster.compute()
print(timedelta(seconds=time()-start))

pd.concat([pd.DataFrame(a[i],columns=['tid','reload_amt','status']) for i in range(len(a))]).to_csv(path+'86 Summary Status.csv')
