==> Development/93 Sensitivity Scenarios for Branch.ipynb <==
import os
import pandas as pd
import numpy as np

import f_optimize as ed

import itertools

from time import time
from datetime import timedelta
pd.options.display.max_rows = 100

import pickle

from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

# # # %load D:/BUDS_MID/__initdask__.py
# # """
# # Created on Wed Feb 19 16:24:03 2020

# # @author: cmchico
# # """
# import dask.bag as db
# from dask.distributed import Client, progress

# client = Client(processes = True,n_workers=12)
# client

path ='D:/BUDS_MID/OJT ATM/ATM Forecasting Data New/'

pd.Series(os.listdir(path+'Prophet Test Perf Thresh R/')).str.split('_').str[1].nunique()

scen=pd.read_csv(path+'93 Scenarios.csv')

scen

# scenario_all.sort_values(by=[' total ','atm_no'],inplace=True)
# scenario_all = scenario_all.reset_index(drop=True)
# scenario_all.head(100)

days_no=7
AVE_AMT = 1000000*np.array(range(4,10))
AVE_AMT

scen.columns

# get_schedule(bb, dbd, dbrel_amt,dbatm_no,days_no=7)
cost=dict()

def errorhandle(b,d,reload_amt,
                 atm_no,days_no,reload_charge=2000, r=0.05/365,t=0,t_day0=0):
    try:
        prob,X = ed.get_schedule(b,d,reload_amt,
                 atm_no,days_no,reload_charge=2000, r=0.05/365,t=0,t_day0=0)
        return prob.solve(),X
    except:
        return 999999999,X

cost=dict()

for amt in AVE_AMT:
    print(amt)
    scenario_all=scen[scen[str(amt)]=="Yes"].copy()
    reload_amt = [[x] for x in scenario_all[scenario_all.atm_no==1].rel1.values] + \
        scenario_all[scenario_all.atm_no==2][['rel1','rel2']].values.tolist() + \
        scenario_all[scenario_all.atm_no==3][['rel1','rel2','rel3']].values.tolist() + \
        scenario_all[scenario_all.atm_no==4][['rel1','rel2','rel3','rel4']].values.tolist()

    datm_no=scenario_all.atm_no.values
    dd = [[amt for k in range(days_no)] for j in range(scenario_all.shape[0])]
    dc = [[0 for k in range(m)] for m in datm_no]

    for i in range(scenario_all.shape[0]):
    # for i in range(2):

        b=dc[i]
        d=dd[i]
        rel_amt=reload_amt[i]
        atm_no=datm_no[i]
        days_no=7
        reload_charge=2000
        r=0.05/365
        t=0
        t_day0=0

        R = np.diag(rel_amt)
        B = np.array([b for i in range(days_no)])
        D = np.array([d])
        T = np.array([[t] for i in range(days_no)])
        if t_day0 >0: T[0] = t_day0
        rel_charge = np.array([[reload_charge] for i in range(atm_no)])

        X=np.ones((days_no,atm_no))
        C = np.tril(np.ones((days_no,days_no)))
        _1 = np.array([np.ones(atm_no)]).T

        civ_cost = r*np.ones((1,days_no))@(B@_1 + C@X@R@_1 - C@D.T)
        reload_cost = np.ones((1,days_no))@X@rel_charge

        tot_cost=civ_cost+reload_cost
        cost[(amt,i)] = [atm_no,civ_cost.flatten()[0],reload_cost.flatten()[0],tot_cost.flatten()[0]]
        print(cost[(amt,i)])

B

C
